shift block setup to parallel.py
