+++ src/mindlessgen/generator/main.py
@@ -10,7 +10,6 @@ from pathlib import Path
 import multiprocessing as mp
 from threading import Event
 import warnings
-from dataclasses import dataclass
 from time import perf_counter
 from datetime import timedelta
 from tqdm import tqdm
@@ -29,7 +28,7 @@ from ..qm import (
     get_gxtb_path,
 )
 from ..molecules import iterative_optimization, postprocess_mol
-from ..prog import ConfigManager, setup_managers, ResourceMonitor
+from ..prog import ConfigManager, setup_managers, ResourceMonitor, setup_blocks
 
 from ..__version__ import __version__
 
@@ -421,50 +420,3 @@ def setup_engines(
         return GXTB(path, cfg.gxtb)
     else:
         raise NotImplementedError("Engine not implemented.")
-
-
-@dataclass
-class Block:
-    num_molecules: int
-    ncores: int
-
-
-def setup_blocks(ncores: int, num_molecules: int, mincores: int) -> list[Block]:
-    blocks: list[Block] = []
-
-    # Maximum and minimum number of parallel processes possible
-    maxcores = ncores
-    maxprocs = max(1, ncores // mincores)
-    minprocs = max(1, ncores // maxcores)
-
-    # Distribute number of molecules among blocks
-    # First (if possible) create the maximum number of parallel blocks (maxprocs) and distribute as many molecules as possible
-    molecules_left = num_molecules
-    if molecules_left >= maxprocs:
-        p = maxprocs
-        molecules_per_block = molecules_left // p
-        for _ in range(p):
-            blocks.append(Block(molecules_per_block, ncores // p))
-        molecules_left -= molecules_per_block * p
-
-    # While there are more than minprocs (1) molecules left find the optimal number of parallel blocks
-    # Again distribute as many molecules per block as possible
-    while molecules_left >= minprocs:
-        p = max(
-            [
-                j
-                for j in range(minprocs, maxprocs)
-                if ncores % j == 0 and j <= molecules_left
-            ]
-        )
-        molecules_per_block = molecules_left // p
-        for _ in range(p):
-            blocks.append(Block(molecules_per_block, ncores // p))
-        molecules_left -= molecules_per_block * p
-
-    # NOTE: using minprocs = 1 this is probably never true
-    if molecules_left > 0:
-        blocks.append(Block(molecules_left, maxcores))
-        molecules_left -= molecules_left
-
-    return blocks
